+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1166,7 +1166,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-12-16T05:41:07.312Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-12-16T10:45:16.987Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1218,7 +1218,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-12-16T05:41:07.311Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-12-16T10:45:16.987Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1270,7 +1270,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[Source] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-12-16T05:41:07.311Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-12-16T10:45:16.986Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1349,5 +1349,5 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-12-16T05:40:51.937Z")
+    updated: Optional[datetime] = parse_datetime("2025-12-16T10:45:03.212Z")
     r"""The date and time the resource was last updated"""