+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -93,5 +93,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-12-16T05:40:51.937Z")
+    updated: Optional[datetime] = parse_datetime("2025-12-16T10:45:03.212Z")
     r"""The date and time the resource was last updated"""