+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -239,7 +239,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KCJTTDGDRF3EEGSDM3ADTAYR"
+        "01KCKC7BEZQCSS93G6GA9TH92T"
     )
     r"""The unique identifier of the data source"""