+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -63,5 +63,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-12-16T05:40:55.480Z")
+    created: Optional[datetime] = parse_datetime("2025-12-16T10:45:06.299Z")
     r"""The date and time the resource was created"""