+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -152,7 +152,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kcjttd2w2qgamm07nz07f9ss"
+        "conv_01kckc7b37pm4xew3swnf4r6x0"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """