+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -130,7 +130,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KCJTTDGBVKYWSDV6QNRG4NWT"
+        "01KCKC7BEXYM0DSKYTK50V4Q0Y"
     )
     r"""The unique identifier of the data source"""