+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KCJTTDGCRKB5Z013Q1PD0C4Z"
+        "01KCKC7BEY8AWRJZDTYMHZ4QVV"
     )
     r"""The unique identifier of the data source"""