+++ packages/orq-rc/src/orq_ai_sdk/models/updatebudgetop.py
@@ -219,5 +219,5 @@ class UpdateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-12-16T05:40:52.364Z")
+    updated: Optional[datetime] = parse_datetime("2025-12-16T10:45:03.543Z")
     r"""The date and time the resource was last updated"""