+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -150,7 +150,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kcjttd30apzza7q57agy6r42"
+        "conv_01kckc7b3abfvzmmtepx5aedtz"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """